实时盈亏增加手牌数

+++ src/main/java/com/dx/facade/texas/resp/DxGameBillResp.java
@@ -90,5 +90,5 @@ public class DxGameBillResp implements Serializable {
     /**
      * Walks手牌数
      */
-    private Long walks;
+    private Integer walks;
 }
